requirements.txt
websocket-client==1.8.0
kafka-python==2.0.2
python-dotenv==1.0.1
fastapi==0.102.0
uvicorn==0.23.2
streamlit==1.30.0
requests==2.31.0
